requirements.txt
# Framework
Django==1.9.6
psycopg2==2.6.1
pytz==2016.4
requests==2.9.1
djangorestframework==3.3.3
django-grappelli==2.8.1
django-countries==3.4.1
django-angular==0.8.2
django-rest-swagger==0.3.6
django-filter==0.13
django-rest-auth==0.7.0
django-allauth==0.25.2
django-ical==1.4

# SSL
cryptography==2.2.2

# Testing
coverage==4.5.1
factory-boy==2.10.0
Faker==0.8.11
mock==2.0.0
requests-mock==1.4.0

# Code Quality
flake8==2.5.4
isort==4.2.5
autopep8==1.2.4
